Cleanup + Kommentare für den Tutor
