--- Blatt4/src/de/buffalodan/ci/network/XORMain.java
@@ -1,9 +1,0 @@
-package de.buffalodan.ci.network;
-
-public class XORMain {
-
-	public static void main(String[] args) {
-		
-	}
-
-}
